refactor(agents-server): deduplicate principal helpers
